feat(readme): add Chinese documentation for Json2Class extension
Added comprehensive documentation in Chinese to the README, detailing the
features, installation, and usage of the Json2Class for VS Code extension.
This includes instructions on how to convert JSON data into Java classes
using the extension and how to contribute to the project.

Also, made a minor adjustment to the .gitignore file to include the .vscode
directory.